perf: Improved the answers, as well as the parameters in the query

+++ controllers/usersTags.controller.js
@@ -1,4 +1,7 @@
-const UsersTagsServices = require("../services/usersTags.service");
+/** @format */
+
+const UsersTagsServices = require('../services/usersTags.service');
+const { CustomError } = require('../utils/helpers');
 const UserTagsService = new UsersTagsServices();
 
 const createUserTag = async (request, response, next) => {
@@ -9,7 +12,7 @@ const createUserTag = async (request, response, next) => {
 
     const tags = await UserTagsService.createTag(body);
     // response.json(vote)
-    response.json({ message: "Interest Added" });
+    response.json({ message: 'Interest Added' });
   } catch (error) {
     next(error);
   }
@@ -17,9 +20,18 @@ const createUserTag = async (request, response, next) => {
 
 const removeUserTag = async (request, response, next) => {
   try {
-    const { tag_id } = request.body;
-    const results = await UserTagsService.removeUserTag(tag_id);
-    response.json({ message: "" });
+    const { isSameUser } = request.user;
+    if (isSameUser) {
+      const { tag_id } = request.body;
+      const results = await UserTagsService.removeUserTag(tag_id);
+      return response.json({ message: 'Interest removed' });
+    }
+
+    throw new CustomError(
+      'You are not authorized to make changes to this user',
+      403,
+      'Unauthorized'
+    );
   } catch (error) {
     next(error);
   }
@@ -27,4 +39,5 @@ const removeUserTag = async (request, response, next) => {
 
 module.exports = {
   createUserTag,
+  removeUserTag,
 };
